seperate out crate for local llm implementation.
