+++ crates/trigger/Cargo.toml
@@ -26,6 +26,7 @@ spin-sqlite-inproc = { path = "../sqlite-inproc" }
 spin-sqlite-libsql = { path = "../sqlite-libsql" }
 spin-world = { path = "../world" }
 spin-llm = { path = "../llm" }
+spin-llm-local = { path = "../llm-local" }
 sanitize-filename = "0.4"
 serde = "1.0"
 serde_json = "1.0"
@@ -49,5 +50,5 @@ tokio = { version = "1.23", features = ["rt", "macros"] }
 
 [features]
 default = []
-llm-metal = ["spin-llm/metal"]
-llm-cublas = ["spin-llm/cublas"]
+llm-metal = ["spin-llm-local/metal"]
+llm-cublas = ["spin-llm-local/cublas"]
