+++ crates/trigger/src/lib.rs
@@ -164,7 +164,7 @@ impl<Executor: TriggerExecutor> TriggerExecutorBuilder<Executor> {
 pub struct HostComponentInitData {
     kv: Vec<(String, String)>,
     sqlite: Vec<String>,
-    llm: spin_llm::LLmOptions,
+    llm: spin_llm_local::LLmOptions,
 }
 
 impl HostComponentInitData {
@@ -174,7 +174,7 @@ impl HostComponentInitData {
     pub fn new(
         key_value_init_values: impl Into<Vec<(String, String)>>,
         sqlite_init_statements: impl Into<Vec<String>>,
-        llm: spin_llm::LLmOptions,
+        llm: spin_llm_local::LLmOptions,
     ) -> Self {
         Self {
             kv: key_value_init_values.into(),
