+++ crates/trigger/src/runtime_config/llm.rs
@@ -1,4 +1,4 @@
-use spin_llm::LlmEngine;
+use spin_llm_local::LocalLlmEngine;
 
 pub(crate) async fn build_component(
     runtime_config: &crate::RuntimeConfig,
@@ -8,7 +8,7 @@ pub(crate) async fn build_component(
         .state_dir()
         .unwrap_or_default()
         .join("ai-models");
-    let mut engine = spin_llm::LocalLlmEngine::new(path, use_gpu);
+    let mut engine = LocalLlmEngine::new(path, use_gpu);
     engine.init().await;
     spin_llm::LlmComponent::new(move || Box::new(engine.clone()))
 }
